dont error out if order.table is null

+++ pos_multi_session_restaurant/static/src/js/pos_multi_session_restaurant.js
@@ -127,7 +127,7 @@ odoo.define('pos_multi_session_restaurant', function(require){
                     return order.uid == data.uid;
                 });
             }
-            if (order && order.table.id != data.table_id) {
+            if (order && order.table && order.table.id != data.table_id) {
                 order.transfer = true;
                 order.destroy({'reason': 'abandon'});
             }
